# Initialize hash conditionally