--- a/lib/optimize/RealContentHashPlugin.js
+++ b/lib/optimize/RealContentHashPlugin.js
@@ -342,7 +342,6 @@ ${referencingAssets
 					for (const oldHash of hashesInOrder) {
 						const assets = hashToAssets.get(oldHash);
 						assets.sort(comparator);
-						const hash = createHash(this._hashFunction);
 						await Promise.all(
 							assets.map(asset =>
 								asset.ownHashes.has(oldHash)
@@ -363,6 +362,7 @@ ${referencingAssets
 						});
 						let newHash = hooks.updateHash.call(assetsContent, oldHash);
 						if (!newHash) {
+							const hash = createHash(this._hashFunction);
 							for (const content of assetsContent) {
 								hash.update(content);
 							}